move new_parse_embl.py to parse_embl.py and update relevant code
