+++ ena_build/__init__.py
@@ -1,4 +1,4 @@
 from . import mysql_database
-from . import new_parse_embl
+from . import parse_embl
 from . import dask_tasks
 from . import dask_tskmgr
